refactor(sdk): Python SDK modular architecture with factory pattern
BREAKING CHANGE: Remove standalone functions in favor of factory pattern.

- Add `create_koine()` factory returning `KoineClient` with methods
- Split monolithic client.py into endpoint modules (text.py, object.py, stream/)
- Expand error codes from 7 to 17 (matching TypeScript SDK)
- Add config validation at client creation time
- Design stream module for future transport extensibility (sockets, websocket)
- Extract internal types to _types.py (not exported)
- Add HTTP utilities module (http.py)
- Split tests by module with shared conftest.py
- Update examples to use factory pattern
- Bump version to 2.0.0

New API:
  koine = create_koine(config)
  result = await koine.generate_text(prompt="Hello")

Removed:
  generate_text(), generate_object(), stream_text() standalone functions
