+++ packages/sdks/python/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "koine-sdk"
-version = "1.1.5"
+version = "2.0.0"
 description = "Python SDK for Koine gateway"
 readme = "README.md"
 license = "MIT"
