+++ packages/sdks/python/src/koine_sdk/stream/__init__.py
@@ -1,0 +1,10 @@
+"""Streaming module for Koine SDK.
+
+This module provides streaming text generation capabilities.
+Currently supports HTTP/SSE transport, with future support planned
+for WebSocket and other transport mechanisms.
+"""
+
+from .http import stream_text
+
+__all__ = ["stream_text"]
